Adding many tests and mocking to check that everything works properly.
The mocking is slightly involved because the Server must use static methods to create dependencies due to the way that Sinatra/Rack works.
Also identified during testing that RestClient does not return a Rack::Response so the client needs to do that translation.
Carried out some refactoring to tidy things up also